+++ Runtime/Components/Cinemachine3rdPersonFollow.cs
@@ -69,21 +69,33 @@ namespace Cinemachine
         /// </summary>
         [Tooltip("Specifies how close the camera can get to obstacles")]
         public float CameraRadius;
+        
+        /// <summary>
+        /// How gradually the camera returns to its normal position after having been corrected by the built-in
+        /// collision resolution system. Higher numbers will move the camera more gradually back to normal.
+        /// </summary>
+        [Range(0, 10)]
+        [Tooltip("How gradually the camera returns to its normal position after having been corrected by the built-in " +
+            "collision resolution system.  Higher numbers will move the camera more gradually back to normal.")]
+        public float CollisionDamping;
 
         // State info
-        private Vector3 PreviousFollowTargetPosition;
-        private float PreviousHeadingAngle;
+        Vector3 m_PreviousFollowTargetPosition;
+        float m_PreviousHeadingAngle;
+        float m_HandCollisionCorrection;
+        float m_CamPosCollisionCorrection;
 
-        private void OnValidate()
+        void OnValidate()
         {
             CameraSide = Mathf.Clamp(CameraSide, -1.0f, 1.0f);
             Damping.x = Mathf.Max(0, Damping.x);
             Damping.y = Mathf.Max(0, Damping.y);
             Damping.z = Mathf.Max(0, Damping.z);
             CameraRadius = Mathf.Max(0.001f, CameraRadius);
+            CollisionDamping = Mathf.Max(0, CollisionDamping);
         }
 
-        private void Reset()
+        void Reset()
         {
             CameraCollisionFilter = 1;
             ShoulderOffset = new Vector3(0.5f, -0.4f, 0.0f);
@@ -92,10 +104,11 @@ namespace Cinemachine
             CameraDistance = 2.0f;
             Damping = new Vector3(0.1f, 0.5f, 0.3f);
             CameraRadius = 0.2f;
+            CollisionDamping = 2f;
         }
 
         /// <summary>True if component is enabled and has a Follow target defined</summary>
-        public override bool IsValid { get { return enabled && FollowTarget != null; } }
+        public override bool IsValid => enabled && FollowTarget != null;
 
         /// <summary>Get the Cinemachine Pipeline stage that this component implements.
         /// Always returns the Aim stage</summary>
@@ -109,24 +122,31 @@ namespace Cinemachine
 
         /// <summary>Orients the camera to match the Follow target's orientation</summary>
         /// <param name="curState">The current camera state</param>
-        /// <param name="deltaTime">Not used.</param>
+        /// <param name="deltaTime">Elapsed time since last frame, for damping calculations.  
+        /// If negative, previous state is reset.</param>
         public override void MutateCameraState(ref CameraState curState, float deltaTime)
         {
-            if (!IsValid)
-                return;
-            if (!VirtualCamera.PreviousStateIsValid)
-                deltaTime = -1;
-            PositionCamera(ref curState, deltaTime);
+            if (IsValid)
+            {
+                if (!VirtualCamera.PreviousStateIsValid)
+                    deltaTime = -1;
+                PositionCamera(ref curState, deltaTime);
+            }
         }
 
         void PositionCamera(ref CameraState curState, float deltaTime)
         {
-            var targetPos = FollowTargetPosition;
-            var prevTargetPos = deltaTime >= 0 ? PreviousFollowTargetPosition : targetPos;
+            var followTarget = FollowTargetPosition;
+            if (deltaTime < 0)
+            {
+                // No damping - reset all state info
+                m_PreviousFollowTargetPosition = followTarget;
+                m_HandCollisionCorrection = m_CamPosCollisionCorrection = 0;
+            }
+            var prevTargetPos = m_PreviousFollowTargetPosition;
 
             // Compute damped target pos (compute in camera space)
-            var dampedTargetPos = Quaternion.Inverse(curState.RawOrientation) 
-                * (targetPos - prevTargetPos);
+            var dampedTargetPos = Quaternion.Inverse(curState.RawOrientation) * (followTarget - prevTargetPos);
             if (deltaTime >= 0)
                 dampedTargetPos = VirtualCamera.DetachedFollowTargetDamp(
                     dampedTargetPos, Damping, deltaTime);
@@ -139,56 +159,81 @@ namespace Cinemachine
             var followTargetForward = followTargetRotation * fwd;
             var angle = UnityVectorExtensions.SignedAngle(
                 fwd, followTargetForward.ProjectOntoPlane(up), up);
-            var previousHeadingAngle = deltaTime >= 0 ? PreviousHeadingAngle : angle;
-            var deltaHeading = angle - previousHeadingAngle;
-            PreviousHeadingAngle = angle;
+            if (deltaTime < 0)
+                m_PreviousHeadingAngle = angle; // no damping
+            var deltaHeading = angle - m_PreviousHeadingAngle;
+            m_PreviousHeadingAngle = angle;
 
             // Bypass user-sourced rotation
-            dampedTargetPos = targetPos 
-                + Quaternion.AngleAxis(deltaHeading, up) * (dampedTargetPos - targetPos);
-            PreviousFollowTargetPosition = dampedTargetPos;
+            dampedTargetPos = followTarget 
+                + Quaternion.AngleAxis(deltaHeading, up) * (dampedTargetPos - followTarget);
+            m_PreviousFollowTargetPosition = dampedTargetPos;
 
-            GetRigPositions(out Vector3 root, out Vector3 shoulder, out Vector3 hand);
+            GetRigPositions(out Vector3 root, out _, out Vector3 hand);
 
             // 1. Check if pivot itself is colliding with something, if yes, then move the pivot 
             // closer to the player. The radius is bigger here than in step 2, to avoid problems 
             // next to walls. Where the preferred distance would be pulled completely to the 
             // player, using a bigger radius, this won't happen.
-            hand = PullTowardsStartOnCollision(in root, in hand, in CameraCollisionFilter, CameraRadius * 1.05f);
+            hand = ResolveCollisions(
+                root, hand, deltaTime, CameraRadius * 1.05f, ref m_HandCollisionCorrection);
 
             // 2. Try to place the camera to the preferred distance
-            var camPos = hand - (followTargetForward * CameraDistance);
-            camPos = PullTowardsStartOnCollision(in hand, in camPos, in CameraCollisionFilter, CameraRadius);
+            Vector3 camPos = hand - (followTargetForward * CameraDistance);
+            camPos = ResolveCollisions(
+                hand, camPos, deltaTime, CameraRadius, ref m_CamPosCollisionCorrection);
 
+            // Set state
             curState.RawPosition = camPos;
-            curState.RawOrientation = FollowTargetRotation;
+            curState.RawOrientation = followTargetRotation;
             curState.ReferenceUp = up;
         }
 
         /// <summary>
         /// Internal use only.  Public for the inspector gizmo
         /// </summary>
-        /// <param name="root"></param>
-        /// <param name="shoulder"></param>
-        /// <param name="hand"></param>
+        /// <param name="root">Root of the rig.</param>
+        /// <param name="shoulder">Shoulder of the rig.</param>
+        /// <param name="hand">Hand of the rig.</param>
         public void GetRigPositions(out Vector3 root, out Vector3 shoulder, out Vector3 hand)
         {
-            root = PreviousFollowTargetPosition;
-            var ShoulderPivotReflected = Vector3.Reflect(ShoulderOffset, Vector3.right);
-            var shoulderOffset = Vector3.Lerp(ShoulderPivotReflected, ShoulderOffset, CameraSide);
-            var handOffset = new Vector3(0, VerticalArmLength, 0);
-            shoulder = root + Quaternion.AngleAxis(PreviousHeadingAngle, Vector3.up) * shoulderOffset;
-            hand = shoulder + FollowTargetRotation * handOffset;
+            root = m_PreviousFollowTargetPosition;
+            var shoulderPivotReflected = Vector3.Reflect(ShoulderOffset, Vector3.right);
+            var shoulderOffset = Vector3.Lerp(shoulderPivotReflected, ShoulderOffset, CameraSide);
+            t_HandOffset.y = VerticalArmLength;
+            shoulder = root + Quaternion.AngleAxis(m_PreviousHeadingAngle, Vector3.up) * shoulderOffset;
+            hand = shoulder + FollowTargetRotation * t_HandOffset;
         }
+        Vector3 t_HandOffset = Vector3.zero; // minor opt. - to avoid creating a new vector in GetRigPositions
 
-        private Vector3 PullTowardsStartOnCollision(
-            in Vector3 rayStart, in Vector3 rayEnd,
-            in LayerMask filter, float radius)
+        Vector3 ResolveCollisions(
+            Vector3 root, Vector3 tip, float deltaTime, 
+            float cameraRadius, ref float collisionCorrection)
         {
-            var dir = rayEnd - rayStart;
+            var dir = tip - root;
+            var len = dir.magnitude;
+            dir /= len;
+
             bool hasHit = RuntimeUtility.SphereCastIgnoreTag(
-                rayStart, radius, dir, out RaycastHit hitInfo, dir.magnitude, filter, IgnoreTag);
-            return hasHit ? hitInfo.point + hitInfo.normal * radius: rayEnd;
+                root, cameraRadius, dir, out RaycastHit hitInfo, 
+                len, CameraCollisionFilter, IgnoreTag);
+
+            var result = tip;
+            if (hasHit)
+            {
+                // We have a collision - store the delta for damping later
+                result = hitInfo.point + hitInfo.normal * cameraRadius;
+                collisionCorrection = (result - tip).magnitude;
+            }
+            else if (deltaTime >= 0)
+            {
+                // Post collision damping - ease out of last collision
+                collisionCorrection -= Damper.Damp(
+                    collisionCorrection, CollisionDamping, deltaTime);
+                if (collisionCorrection > Epsilon)
+                    result -= dir * collisionCorrection;
+            }
+            return result;
         }
     }
 #endif